// src/API.ts
export type Post = {
  __typename: "Post",
  id: string,
  title: string,
  content: string,
  image?: string | null,
  votes?: ModelVoteConnection | null,
  comments?: ModelCommentConnection | null,
  createdAt: string,
  updatedAt: string,
  owner?: string | null,
};

export type ModelVoteConnection = {
  __typename: "ModelVoteConnection",
  items?:  Array<Vote | null >,
  nextToken?: string | null,
};

export type Vote = {
  __typename: "Vote",
  id: string,
  postID: string,
  vote: string,
  createdAt: string,
  updatedAt: string,
  owner?: string | null,
};

export type ModelCommentConnection = {
  __typename: "ModelCommentConnection",
  items?:  Array<Comment | null >,
  nextToken?: string | null,
};

export type Comment = {
  __typename: "Comment",
  id: string,
  postID: string,
  content: string,
  createdAt: string,
  updatedAt: string,
  owner?: string | null,
};
// ...
export type CreateVoteInput = {
  id?: string | null,
  postID: string,
  vote: string,
};

export type ModelVoteConditionInput = {
  postID?: ModelIDInput | null,
  vote?: ModelStringInput | null,
  and?: Array< ModelVoteConditionInput | null > | null,
  or?: Array< ModelVoteConditionInput | null > | null,
  not?: ModelVoteConditionInput | null,
};

export type UpdateVoteInput = {
  id: string,
  postID?: string | null,
  vote?: string | null,
};
// ...
export type CreateVoteMutation = {
  createVote?:  {
    __typename: "Vote",
    id: string,
    postID: string,
    vote: string,
    createdAt: string,
    updatedAt: string,
    owner?: string | null,
  } | null,
};

export type UpdateVoteMutationVariables = {
  input: UpdateVoteInput,
  condition?: ModelVoteConditionInput | null,
};

export type UpdateVoteMutation = {
  updateVote?:  {
    __typename: "Vote",
    id: string,
    postID: string,
    vote: string,
    createdAt: string,
    updatedAt: string,
    owner?: string | null,
  } | null,
};